requirements.txt
imgaug
pillow
opencv-python
numpy